refactor: #3 request domain
