+++ src/main/java/com/yourmode/yourmodebackend/domain/request/dto/ContentRequestStatusHistoryDto.java
@@ -4,7 +4,6 @@ import lombok.*;
 import java.time.LocalDateTime;
 
 @Getter
-@Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
