+++ src/main/java/com/yourmode/yourmodebackend/domain/request/dto/editor/response/EditorContentRequestDetailDto.java
@@ -3,27 +3,26 @@ package com.yourmode.yourmodebackend.domain.request.dto.editor.response;
 import com.yourmode.yourmodebackend.domain.request.dto.UserProfileSummaryDto;
 import com.yourmode.yourmodebackend.domain.user.entity.User;
 import lombok.Getter;
-import lombok.Setter;
+import lombok.Builder;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Getter
-@Setter
+@Builder
 public class EditorContentRequestDetailDto {
-    private Long id;
-    private User user;
-    private UserProfileSummaryDto profile;
-    private String bodyFeature;
-    private String situation;
-    private String recommendedStyle;
-    private String avoidedStyle;
-    private Integer budget;
-    private Boolean isPublic;
-    private String status;
+    private final Long id;
+    private final UserProfileSummaryDto profile;
+    private final String bodyFeature;
+    private final String situation;
+    private final String recommendedStyle;
+    private final String avoidedStyle;
+    private final Integer budget;
+    private final Boolean isPublic;
+    private final String status;
 
-    private List<Long> itemCategoryIds;       // 기존 유지 (필요 시)
-    private List<String> itemCategoryNames;
+    private final List<Long> itemCategoryIds;       // 기존 유지 (필요 시)
+    private final List<String> itemCategoryNames;
 
-    private LocalDateTime createdAt;
+    private final LocalDateTime createdAt;
 }
