+++ src/main/java/com/yourmode/yourmodebackend/domain/request/dto/editor/response/EditorContentRequestSummaryDto.java
@@ -11,16 +11,16 @@ import java.util.List;
 @Getter
 @Builder
 public class EditorContentRequestSummaryDto {
-    private UserProfileSummaryDto profile;
-    private Long id;
-    private String bodyFeature;
-    private String situation;
-    private String recommendedStyle;
+    private final UserProfileSummaryDto profile;
+    private final Long id;
+    private final String bodyFeature;
+    private final String situation;
+    private final String recommendedStyle;
 
-    private List<Long> itemCategoryIds;       // 기존 유지 (필요 시)
-    private List<String> itemCategoryNames;
+    private final List<Long> itemCategoryIds;       // 기존 유지 (필요 시)
+    private final List<String> itemCategoryNames;
 
-    private LocalDateTime createdAt;
+    private final LocalDateTime createdAt;
 
-    private List<ContentRequestStatusHistoryDto> statusHistories; // 상태 변경 이력
+    private final List<ContentRequestStatusHistoryDto> statusHistories; // 상태 변경 이력
 }
