+++ src/main/java/com/yourmode/yourmodebackend/domain/request/dto/user/response/UserContentRequestDetailDto.java
@@ -7,7 +7,6 @@ import lombok.*;
 import java.util.List;
 
 @Getter
-@Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
