+++ src/main/java/com/yourmode/yourmodebackend/domain/request/dto/user/response/UserContentRequestSummaryDto.java
@@ -6,7 +6,6 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Getter
-@Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
